(wip) : uplifting states to hooks
