--- src/hooks/useCustomHook.ts
@@ -1,4 +1,0 @@
-//  This is just a placeholder file
-export const useCustomHook = (): string => {
-  return 'This is Custom hook '
-}
