+++ src/pages/Styles.ts
@@ -1,0 +1,12 @@
+import styled from 'styled-components/macro'
+
+export const Container = styled.div`
+  display: flex;
+  flex-grow: 1;
+  flex-direction: column;
+`
+
+export const InputContainer = styled.div`
+  margin: 8px;
+  margin: 0 auto;
+`
